requirements.txt
jupyter==1.0.0
Pillow==9.4.0
numpy==2.3.4
pandas==2.3.3
matplotlib==3.10.7
seaborn==0.13.2
torch==2.9.1
torchvision==0.24.1